LPS-179975 Refectoring Claim Details Page
