+++ modules/apps/site-initializer/site-initializer-raylife-ap/extra/remote-app/src/common/components/activities-table-list/index.tsx
@@ -45,7 +45,7 @@ const HEADERSMOBILE: TableListHeaders[] = [
 ];
 
 type ActivitieDataType = {
-	BodyElement?: any;
+	BodyElement?: () => void;
 	activitiesData: {[keys: string]: string | boolean}[];
 };
