+++ modules/apps/site-initializer/site-initializer-raylife-ap/extra/remote-app/src/routes/claims/pages/ClaimsDetails/Types/index.ts
@@ -54,3 +54,8 @@ export type ClaimActivitiesDataType = {
 	date: string;
 	message: string;
 };
+
+export type ClaimComponentsType = {
+	claimData: ClaimType;
+	isClaimSettled?: boolean;
+};
