+++ modules/apps/site-initializer/site-initializer-raylife-ap/extra/remote-app/src/routes/claims/pages/ClaimsDetails/claim-action-details/index.scss
@@ -1,0 +1,13 @@
+.action-detail-content {
+	.lexicon-icon {
+		font-size: 12px !important;
+	}
+
+	.cursor-pointer {
+		cursor: pointer;
+
+		&:hover {
+			text-decoration: underline;
+		}
+	}
+}
