+++ modules/apps/site-initializer/site-initializer-raylife-ap/extra/remote-app/src/routes/claims/pages/ClaimsDetails/claim-action-details/notify-repair-component/index.tsx
